[#69] Rename traits into imperative form and impl back to original version
