+++ iceoryx2/src/lib.rs
@@ -288,16 +288,16 @@ pub mod port;
 pub(crate) mod raw_sample;
 
 /// The interface for the payload that is received by a [`crate::port::subscriber::Subscriber`].
-pub mod sample;
+pub mod payload;
 
 /// The payload that is received by a [`crate::port::subscriber::Subscriber`].
-pub mod sample_impl;
+pub mod sample;
 
 /// The interface for the payload that is sent by a [`crate::port::publisher::Publisher`].
-pub mod sample_mut;
+pub mod payload_mut;
 
 /// The payload that is sent by a [`crate::port::publisher::Publisher`].
-pub mod sample_mut_impl;
+pub mod sample_mut;
 
 /// The foundation of communication the service with its
 /// [`MessagingPattern`](crate::service::messaging_pattern::MessagingPattern)
