+++ iceoryx2/src/port/listener.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2023 - 2024 Contributors to the Eclipse Foundation
+// Copyright (c) 2023 Contributors to the Eclipse Foundation
 //
 // See the NOTICE file(s) distributed with this work for additional
 // information regarding copyright ownership.
@@ -14,78 +14,143 @@
 //!
 //! ```
 //! use iceoryx2::prelude::*;
-//! use std::boxed::Box;
 //! # fn main() -> Result<(), Box<dyn std::error::Error>> {
 //! let event_name = ServiceName::new("MyEventName")?;
-//! let event_ipc = zero_copy::Service::new(&event_name)
+//! let event = zero_copy::Service::new(&event_name)
 //!     .event()
 //!     .open_or_create()?;
 //!
-//! let event_local = process_local::Service::new(&event_name)
-//!     .event()
-//!     .open_or_create()?;
+//! let mut listener = event.listener().create()?;
 //!
-//! let mut listeners: Vec<Box<dyn Listener>> = vec![];
-//!
-//! listeners.push(Box::new(event_ipc.listener().create()?));
-//! listeners.push(Box::new(event_local.listener().create()?));
-//!
-//! for listener in &mut listeners {
-//!     for event_id in listener.try_wait()? {
-//!         println!("event was triggered with id: {:?}", event_id);
-//!     }
+//! for event_id in listener.try_wait()? {
+//!     println!("event was triggered with id: {:?}", event_id);
 //! }
 //!
 //! # Ok(())
 //! # }
 //! ```
 //!
-//! See also [`crate::port::listener_impl::ListenerImpl`]
+//! See also [`crate::port::listener::Listener`]
 
-use std::time::Duration;
+use iceoryx2_bb_lock_free::mpmc::unique_index_set::UniqueIndex;
+use iceoryx2_bb_log::fail;
+use iceoryx2_cal::dynamic_storage::DynamicStorage;
+use iceoryx2_cal::event::{ListenerBuilder, ListenerWaitError};
+use iceoryx2_cal::named_concept::NamedConceptBuilder;
 
-use iceoryx2_cal::event::ListenerWaitError;
+use crate::service::naming_scheme::event_concept_name;
+use crate::{port::port_identifiers::UniqueListenerId, service};
+use std::{marker::PhantomData, time::Duration};
 
 use super::event_id::EventId;
+use super::listen::{Listen, ListenerCreateError};
 
-/// Defines the failures that can occur when a [`Listener`] is created with the
-/// [`crate::service::port_factory::listener::PortFactoryListener`].
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum ListenerCreateError {
-    ExceedsMaxSupportedListeners,
-    ResourceCreationFailed,
+/// Represents the receiving endpoint of an event based communication.
+#[derive(Debug)]
+pub struct Listener<'a, 'config: 'a, Service: service::Details<'config>> {
+    _dynamic_config_guard: Option<UniqueIndex<'a>>,
+    listener: <Service::Event as iceoryx2_cal::event::Event<EventId>>::Listener,
+    cache: Vec<EventId>,
+    _phantom_a: PhantomData<&'a Service>,
+    _phantom_b: PhantomData<&'config ()>,
 }
 
-impl std::fmt::Display for ListenerCreateError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        std::write!(f, "{}::{:?}", std::stringify!(Self), self)
+impl<'a, 'config: 'a, Service: service::Details<'config>> Listener<'a, 'config, Service> {
+    pub(crate) fn new(service: &'a Service) -> Result<Self, ListenerCreateError> {
+        let msg = "Failed to create listener";
+        let origin = "Listener::new()";
+        let port_id = UniqueListenerId::new();
+
+        let event_name = event_concept_name(&port_id);
+        let listener = fail!(from origin,
+                             when <Service::Event as iceoryx2_cal::event::Event<EventId>>::ListenerBuilder::new(&event_name).create(),
+                             with ListenerCreateError::ResourceCreationFailed,
+                             "{} since the underlying event concept \"{}\" could not be created.", msg, event_name);
+
+        let mut new_self = Self {
+            _dynamic_config_guard: None,
+            listener,
+            cache: vec![],
+            _phantom_a: PhantomData,
+            _phantom_b: PhantomData,
+        };
+
+        // !MUST! be the last task otherwise a listener is added to the dynamic config without
+        // the creation of all required channels
+        new_self._dynamic_config_guard = Some(
+            match service
+                .state()
+                .dynamic_storage
+                .get()
+                .event()
+                .add_listener_id(port_id)
+            {
+                Some(unique_index) => unique_index,
+                None => {
+                    fail!(from origin, with ListenerCreateError::ExceedsMaxSupportedListeners,
+                                 "{} since it would exceed the maximum supported amount of listeners of {}.",
+                                 msg, service.state().static_config.event().max_listeners);
+                }
+            },
+        );
+
+        Ok(new_self)
+    }
+
+    fn fill_cache(&mut self) -> Result<(), ListenerWaitError> {
+        use iceoryx2_cal::event::Listener;
+        while let Some(id) = fail!(from self,
+                when self.listener.try_wait(),
+                "Failed to try_wait on Listener port since the underlying Listener concept failed.")
+        {
+            self.cache.push(id);
+        }
+
+        Ok(())
     }
 }
 
-impl std::error::Error for ListenerCreateError {}
+impl<'a, 'config: 'a, Service: service::Details<'config>> Listen
+    for Listener<'a, 'config, Service>
+{
+    fn cache(&self) -> &[EventId] {
+        &self.cache
+    }
 
-/// The interface of the receiving endpoint of an event based communication.
-pub trait Listener {
-    /// Returns the cached [`EventId`]s. Whenever [`Listener::try_wait()`],
-    /// [`Listener::timed_wait()`] or [`Listener::blocking_wait()`] is called the cache is reset
-    /// and filled with the events that where signaled since the last call. This cache can be
-    /// accessed until a new wait call resets and fills it again.
-    fn cache(&self) -> &[EventId];
+    fn try_wait(&mut self) -> Result<&[EventId], ListenerWaitError> {
+        self.cache.clear();
+        self.fill_cache()?;
 
-    /// Non-blocking wait for new [`EventId`]s. If no [`EventId`]s were notified the returned slice
-    /// is empty. On error it returns [`ListenerWaitError`] is returned which describes the error
-    /// in detail.
-    fn try_wait(&mut self) -> Result<&[EventId], ListenerWaitError>;
+        Ok(self.cache())
+    }
 
-    /// Blocking wait for new [`EventId`]s until either an [`EventId`] was received or the timeout
-    /// has passed. If no [`EventId`]s were notified the returned slice
-    /// is empty. On error it returns [`ListenerWaitError`] is returned which describes the error
-    /// in detail.
-    fn timed_wait(&mut self, timeout: Duration) -> Result<&[EventId], ListenerWaitError>;
+    fn timed_wait(&mut self, timeout: Duration) -> Result<&[EventId], ListenerWaitError> {
+        use iceoryx2_cal::event::Listener;
+        self.cache.clear();
 
-    /// Blocking wait for new [`EventId`]s until either an [`EventId`].
-    /// Sporadic wakeups can occur and if no [`EventId`]s were notified the returned slice
-    /// is empty. On error it returns [`ListenerWaitError`] is returned which describes the error
-    /// in detail.
-    fn blocking_wait(&mut self) -> Result<&[EventId], ListenerWaitError>;
+        if let Some(id) = fail!(from self,
+            when self.listener.timed_wait(timeout),
+            "Failed to timed_wait with timeout {:?} on Listener port since the underlying Listener concept failed.", timeout)
+        {
+            self.cache.push(id);
+            self.fill_cache()?;
+        }
+
+        Ok(self.cache())
+    }
+
+    fn blocking_wait(&mut self) -> Result<&[EventId], ListenerWaitError> {
+        use iceoryx2_cal::event::Listener;
+        self.cache.clear();
+
+        if let Some(id) = fail!(from self,
+            when self.listener.blocking_wait(),
+            "Failed to blocking_wait on Listener port since the underlying Listener concept failed.")
+        {
+            self.cache.push(id);
+            self.fill_cache()?;
+        }
+
+        Ok(self.cache())
+    }
 }
