+++ iceoryx2/src/prelude.rs
@@ -12,10 +12,10 @@
 
 pub use crate::iox2::Iox2;
 pub use crate::iox2::Iox2Event;
+pub use crate::payload::Payload;
+pub use crate::payload_mut::{PayloadMut, UninitializedPayloadMut};
 pub use crate::port::{
-    event_id::EventId, listener::Listener, notifier::Notifier, publisher::Publisher,
-    publisher::PublisherLoan, subscriber::Subscriber,
+    event_id::EventId, listen::Listen, notify::Notify, publish::Publish, publish::PublishLoan,
+    subscribe::Subscribe,
 };
-pub use crate::sample::Sample;
-pub use crate::sample_mut::{SampleMut, UninitializedSampleMut};
 pub use crate::service::{process_local, service_name::ServiceName, zero_copy, Details, Service};
