+++ iceoryx2/src/sample.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Contributors to the Eclipse Foundation
+// Copyright (c) 2023 Contributors to the Eclipse Foundation
 //
 // See the NOTICE file(s) distributed with this work for additional
 // information regarding copyright ownership.
@@ -21,30 +21,54 @@
 //! #   .open_or_create::<u64>()?;
 //! # let subscriber = service.subscriber().create()?;
 //!
-//! let mut received_samples: Vec<Box<dyn Sample<u64>>> = vec![];
-//!
 //! while let Some(sample) = subscriber.receive()? {
 //!     println!("received: {:?}", *sample);
 //!     println!("header timestamp {:?}, publisher id {:?}",
 //!         sample.header().time_stamp(), sample.header().publisher_id());
-//!     received_samples.push(Box::new(sample));
 //! }
 //!
 //! # Ok(())
 //! # }
 //! ```
 //!
-//! See also [`crate::sample_impl::SampleImpl`].
+//! See also [`crate::sample::Sample`].
 
+use std::{fmt::Debug, ops::Deref};
+
+use crate::port::subscribe::internal::SubscribeMgmt;
 use crate::service::header::publish_subscribe::Header;
+use crate::{payload::Payload, raw_sample::RawSample};
 
 /// It stores the payload and is acquired by the [`crate::port::subscriber::Subscriber`] whenever
 /// it receives new data from a [`crate::port::publisher::Publisher`] via
 /// [`crate::port::subscriber::Subscriber::receive()`].
-pub trait Sample<MessageType> {
-    /// Returns a reference to the payload of the sample
-    fn payload(&self) -> &MessageType;
+#[derive(Debug)]
+pub struct Sample<'subscriber, MessageType: Debug> {
+    pub(crate) subscriber: &'subscriber dyn SubscribeMgmt,
+    pub(crate) ptr: RawSample<Header, MessageType>,
+    pub(crate) channel_id: usize,
+}
 
-    /// Returns a reference to the header of the sample.
-    fn header(&self) -> &Header;
+impl<MessageType: Debug> Deref for Sample<'_, MessageType> {
+    type Target = MessageType;
+    fn deref(&self) -> &Self::Target {
+        self.ptr.as_data_ref()
+    }
+}
+
+impl<MessageType: Debug> Drop for Sample<'_, MessageType> {
+    fn drop(&mut self) {
+        self.subscriber
+            .release_sample(self.channel_id, self.ptr.as_ptr() as usize);
+    }
+}
+
+impl<'subscriber, MessageType: Debug> Payload<MessageType> for Sample<'subscriber, MessageType> {
+    fn payload(&self) -> &MessageType {
+        self.ptr.as_data_ref()
+    }
+
+    fn header(&self) -> &Header {
+        self.ptr.as_header_ref()
+    }
 }
