+++ iceoryx2/src/sample_mut.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Contributors to the Eclipse Foundation
+// Copyright (c) 2023 Contributors to the Eclipse Foundation
 //
 // See the NOTICE file(s) distributed with this work for additional
 // information regarding copyright ownership.
@@ -12,205 +12,123 @@
 
 //! # Example
 //!
-//! ## Sample Write Function
-//!
 //! ```
 //! use iceoryx2::prelude::*;
-//!
-//! fn write_sample_data<S: UninitializedSampleMut<u64>>(mut sample: S) -> impl SampleMut<u64> {
-//!     sample.write_payload(123456)
-//! }
-//!
 //! # fn main() -> Result<(), Box<dyn std::error::Error>> {
 //! # let service_name = ServiceName::new("My/Funk/ServiceName").unwrap();
 //! #
-//! let service = zero_copy::Service::new(&service_name)
-//!     .publish_subscribe()
-//!     .open_or_create::<u64>()?;
-//!
-//! let publisher = service.publisher().create()?;
+//! # let service = zero_copy::Service::new(&service_name)
+//! #     .publish_subscribe()
+//! #     .open_or_create::<u64>()?;
+//! #
+//! # let publisher = service.publisher().create()?;
 //!
 //! let sample = publisher.loan_uninit()?;
-//! let sample = write_sample_data(sample);
+//! let sample = sample.write_payload(1234);
 //!
+//! println!("timestamp: {:?}, publisher port id: {:?}",
+//!     sample.header().time_stamp(), sample.header().publisher_id());
 //! sample.send()?;
 //!
 //! # Ok(())
 //! # }
 //! ```
-//! See also, [`crate::sample_mut_impl::SampleMutImpl`].
+//!
+//! See also, [`crate::sample_mut::SampleMut`].
 
-use iceoryx2_cal::zero_copy_connection::ZeroCopyCreationError;
-
-use crate::service::header::publish_subscribe::Header;
-
-pub(crate) mod internal {
-    use iceoryx2_cal::zero_copy_connection::PointerOffset;
-
-    pub trait SampleMgmt {
-        fn offset_to_chunk(&self) -> PointerOffset;
-    }
-}
+use crate::{
+    payload_mut::{internal::PayloadMgmt, PayloadMut, UninitializedPayloadMut},
+    port::publish::internal::PublishMgmt,
+    raw_sample::RawSampleMut,
+    service::header::publish_subscribe::Header,
+};
+use iceoryx2_cal::shared_memory::*;
+use std::{fmt::Debug, mem::MaybeUninit};
 
 /// Acquired by a [`crate::port::publisher::Publisher`] via
-/// [`crate::port::publisher::PublisherLoan::loan()`]. It stores the payload that will be sent
-/// to all connected [`crate::port::subscriber::Subscriber`]s. If the [`SampleMut`] is not sent
-/// it will release the loaned memory when going out of scope.
-pub trait SampleMut<MessageType>: internal::SampleMgmt {
-    /// Returns a reference to the header of the sample.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let service_name = ServiceName::new("My/Funk/ServiceName").unwrap();
-    /// #
-    /// # let service = zero_copy::Service::new(&service_name)
-    /// #     .publish_subscribe()
-    /// #     .open_or_create::<u64>()?;
-    /// # let publisher = service.publisher().create()?;
-    ///
-    /// let sample = publisher.loan()?;
-    /// println!("Sample Publisher Origin {:?}", sample.header().publisher_id());
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    fn header(&self) -> &Header;
-
-    /// Returns a reference to the payload of the sample.
-    ///
-    /// # Notes
-    ///
-    /// The generic parameter `MessageType` can be packed into [`core::mem::MaybeUninit<MessageType>`], depending
-    /// which API is used to obtain the sample. Obtaining a reference is safe for either type.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let service_name = ServiceName::new("My/Funk/ServiceName").unwrap();
-    /// #
-    /// # let service = zero_copy::Service::new(&service_name)
-    /// #     .publish_subscribe()
-    /// #     .open_or_create::<u64>()?;
-    /// # let publisher = service.publisher().create()?;
-    ///
-    /// let sample = publisher.loan()?;
-    /// println!("Sample current payload {}", sample.payload());
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    fn payload(&self) -> &MessageType;
-
-    /// Returns a mutable reference to the payload of the sample.
-    ///
-    /// # Notes
-    ///
-    /// The generic parameter `MessageType` can be packed into [`core::mem::MaybeUninit<MessageType>`], depending
-    /// which API is used to obtain the sample. Obtaining a reference is safe for either type.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let service_name = ServiceName::new("My/Funk/ServiceName").unwrap();
-    /// #
-    /// # let service = zero_copy::Service::new(&service_name)
-    /// #     .publish_subscribe()
-    /// #     .open_or_create::<u64>()?;
-    /// # let publisher = service.publisher().create()?;
-    ///
-    /// let mut sample = publisher.loan()?;
-    /// *sample.payload_mut() = 4567;
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    fn payload_mut(&mut self) -> &mut MessageType;
-
-    /// Send a previously loaned [`crate::port::publisher::Publisher::loan_uninit()`] or
-    /// [`crate::port::publisher::PublisherLoan::loan()`] [`SampleMut`] to all connected
-    /// [`crate::port::subscriber::Subscriber`]s of the service.
-    ///
-    /// The payload of the [`SampleMut`] must be initialized before it can be sent. Have a look
-    /// at [`UninitializedSampleMut::write_payload()`] and [`UninitializedSampleMut::assume_init()`]
-    /// for more details.
-    ///
-    /// On success the number of [`crate::port::subscriber::Subscriber`]s that received
-    /// the data is returned, otherwise a [`ZeroCopyCreationError`] describing the failure.
-    fn send(self) -> Result<usize, ZeroCopyCreationError>;
-}
-
-/// Acquired by a [`crate::port::publisher::Publisher`] via
+/// [`crate::port::publisher::PublisherLoan::loan()`] or
 /// [`crate::port::publisher::Publisher::loan_uninit()`]. It stores the payload that will be sent
 /// to all connected [`crate::port::subscriber::Subscriber`]s. If the [`SampleMut`] is not sent
 /// it will release the loaned memory when going out of scope.
 ///
-/// The generic parameter `MessageType` is packed into [`core::mem::MaybeUninit<MessageType>`]
-pub trait UninitializedSampleMut<MessageType> {
-    type InitializedSample: SampleMut<MessageType>;
+/// # Notes
+///
+/// Does not implement [`Send`] since it releases unsent samples in the [`crate::port::publisher::Publisher`] and the
+/// [`crate::port::publisher::Publisher`] is not thread-safe!
+///
+/// The generic parameter `M` is either a `MessageType` or a [`core::mem::MaybeUninit<MessageType>`], depending
+/// which API is used to obtain the sample.
+#[derive(Debug)]
+pub struct SampleMut<'publisher, M: Debug> {
+    pub(crate) publisher: &'publisher dyn PublishMgmt,
+    ptr: RawSampleMut<Header, M>,
+    offset_to_chunk: PointerOffset,
+}
 
-    /// Writes the payload to the sample and labels the sample as initialized
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let service_name = ServiceName::new("My/Funk/ServiceName").unwrap();
-    /// #
-    /// # let service = zero_copy::Service::new(&service_name)
-    /// #     .publish_subscribe()
-    /// #     .open_or_create::<u64>()?;
-    /// #
-    /// # let publisher = service.publisher().create()?;
-    ///
-    /// let sample = publisher.loan_uninit()?;
-    /// let sample = sample.write_payload(1234);
-    ///
-    /// sample.send()?;
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    fn write_payload(self, value: MessageType) -> Self::InitializedSample;
+impl<M: Debug> Drop for SampleMut<'_, M> {
+    fn drop(&mut self) {
+        self.publisher.return_loaned_sample(self.offset_to_chunk);
+    }
+}
 
-    /// Extracts the value of the [`core::mem::MaybeUninit<MessageType>`] container and labels the sample as initialized
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that [`core::mem::MaybeUninit<MessageType>`] really is initialized. Calling this when
-    /// the content is not fully initialized causes immediate undefined behavior.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let service_name = ServiceName::new("My/Funk/ServiceName").unwrap();
-    /// #
-    /// # let service = zero_copy::Service::new(&service_name)
-    /// #     .publish_subscribe()
-    /// #     .open_or_create::<u64>()?;
-    /// #
-    /// # let publisher = service.publisher().create()?;
-    ///
-    /// let mut sample = publisher.loan_uninit()?;
-    /// sample.payload_mut().write(1234);
-    /// let sample = unsafe { sample.assume_init() };
-    ///
-    /// sample.send()?;
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    unsafe fn assume_init(self) -> Self::InitializedSample;
+impl<'publisher, MessageType: Debug> SampleMut<'publisher, MaybeUninit<MessageType>> {
+    pub(crate) fn new(
+        publisher: &'publisher dyn PublishMgmt,
+        ptr: RawSampleMut<Header, MaybeUninit<MessageType>>,
+        offset_to_chunk: PointerOffset,
+    ) -> Self {
+        // SAFETY: the transmute is not nice but safe since MaybeUninit is #[repr(transparent)} to the inner type
+        let publisher = unsafe { std::mem::transmute(publisher) };
+
+        Self {
+            publisher,
+            ptr,
+            offset_to_chunk,
+        }
+    }
+}
+
+impl<'publisher, MessageType: Debug> PayloadMgmt for SampleMut<'publisher, MessageType> {
+    fn offset_to_chunk(&self) -> PointerOffset {
+        self.offset_to_chunk
+    }
+}
+
+impl<'publisher, MessageType: Debug> UninitializedPayloadMut<MessageType>
+    for SampleMut<'publisher, MaybeUninit<MessageType>>
+{
+    type InitializedSample = SampleMut<'publisher, MessageType>;
+
+    fn write_payload(mut self, value: MessageType) -> SampleMut<'publisher, MessageType> {
+        self.payload_mut().write(value);
+        // SAFETY: this is safe since the payload was initialized on the line above
+        unsafe { self.assume_init() }
+    }
+
+    unsafe fn assume_init(self) -> SampleMut<'publisher, MessageType> {
+        // the transmute is not nice but safe since MaybeUninit is #[repr(transparent)] to the inner type
+        std::mem::transmute(self)
+    }
+}
+
+impl<
+        'publisher,
+        M: Debug, // `M` is either a `MessageType` or a `MaybeUninit<MessageType>`
+    > PayloadMut<M> for SampleMut<'publisher, M>
+{
+    fn header(&self) -> &Header {
+        self.ptr.as_header_ref()
+    }
+
+    fn payload(&self) -> &M {
+        self.ptr.as_data_ref()
+    }
+
+    fn payload_mut(&mut self) -> &mut M {
+        self.ptr.as_data_mut()
+    }
+
+    fn send(self) -> Result<usize, iceoryx2_cal::zero_copy_connection::ZeroCopyCreationError> {
+        self.publisher.send_impl(self.offset_to_chunk.value())
+    }
 }
