+++ iceoryx2/src/service/port_factory/publisher.rs
@@ -36,9 +36,9 @@ use iceoryx2_bb_log::fail;
 use serde::{de::Visitor, Deserialize, Serialize};
 
 use super::publish_subscribe::PortFactory;
-use crate::{port::publisher::PublisherCreateError, port::publisher_impl::PublisherImpl, service};
+use crate::{port::publish::PublisherCreateError, port::publisher::Publisher, service};
 
-/// Defines the strategy the [`PublisherImpl`] shall pursue in
+/// Defines the strategy the [`Publisher`] shall pursue in
 /// [`crate::sample_mut::SampleMut::send()`] or
 /// [`crate::port::publisher::Publisher::send_copy()`] when the buffer of a
 /// [`crate::port::subscriber::Subscriber`] is full and the service does not overflow.
@@ -99,7 +99,7 @@ pub(crate) struct LocalPublisherConfig {
     pub(crate) unable_to_deliver_strategy: UnableToDeliverStrategy,
 }
 
-/// Factory to create a new [`PublisherImpl`] port/endpoint for
+/// Factory to create a new [`Publisher`] port/endpoint for
 /// [`MessagingPattern::PublishSubscribe`](crate::service::messaging_pattern::MessagingPattern::PublishSubscribe) based
 /// communication.
 #[derive(Debug)]
@@ -138,7 +138,7 @@ impl<'factory, 'config, Service: service::Details<'config>, MessageType: Debug>
         }
     }
 
-    /// Defines how many [`crate::sample_mut::SampleMut`] the [`PublisherImpl`] can loan with
+    /// Defines how many [`crate::sample_mut::SampleMut`] the [`Publisher`] can loan with
     /// [`crate::port::publisher::PublisherLoan::loan()`] or
     /// [`crate::port::publisher::Publisher::loan_uninit()`] in parallel.
     pub fn max_loaned_samples(mut self, value: usize) -> Self {
@@ -152,12 +152,12 @@ impl<'factory, 'config, Service: service::Details<'config>, MessageType: Debug>
         self
     }
 
-    /// Creates a new [`PublisherImpl`] or returns a [`PublisherCreateError`] on failure.
+    /// Creates a new [`Publisher`] or returns a [`PublisherCreateError`] on failure.
     pub fn create(
         self,
-    ) -> Result<PublisherImpl<'factory, 'config, Service, MessageType>, PublisherCreateError> {
+    ) -> Result<Publisher<'factory, 'config, Service, MessageType>, PublisherCreateError> {
         Ok(
-            fail!(from self, when PublisherImpl::new(&self.factory.service, self.factory.service.state().static_config.publish_subscribe(), &self.config),
+            fail!(from self, when Publisher::new(&self.factory.service, self.factory.service.state().static_config.publish_subscribe(), &self.config),
                 "Failed to create new Publisher port."),
         )
     }
