+++ iceoryx2/tests/publisher_tests.rs
@@ -14,9 +14,9 @@
 mod publisher {
     use std::time::{Duration, Instant};
 
-    use iceoryx2::port::publisher::PublisherLoanError;
+    use iceoryx2::payload_mut::UninitializedPayloadMut;
+    use iceoryx2::port::publish::PublisherLoanError;
     use iceoryx2::prelude::*;
-    use iceoryx2::sample_mut::UninitializedSampleMut;
     use iceoryx2::service::port_factory::publisher::UnableToDeliverStrategy;
     use iceoryx2::service::{service_name::ServiceName, Service};
     use iceoryx2_bb_posix::barrier::{BarrierBuilder, BarrierHandle};
