+++ iceoryx2/tests/service_publish_subscribe_tests.rs
@@ -13,8 +13,8 @@
 #[generic_tests::define]
 mod service_publish_subscribe {
     use iceoryx2::config::Config;
-    use iceoryx2::port::publisher::{PublisherCreateError, PublisherLoanError};
-    use iceoryx2::port::subscriber::SubscriberCreateError;
+    use iceoryx2::port::publish::{PublisherCreateError, PublisherLoanError};
+    use iceoryx2::port::subscribe::SubscriberCreateError;
     use iceoryx2::prelude::*;
     use iceoryx2::service::builder::publish_subscribe::PublishSubscribeCreateError;
     use iceoryx2::service::builder::publish_subscribe::PublishSubscribeOpenError;
